requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
gunicorn==25.3.0
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
